pyproject.toml
[project]
name = "rag-agent-merged"
version = "1.0.0"
description = "Agentic RAG Pipeline — Chainlit UI + LangGraph + Qdrant Hybrid + MCP + Reranking"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    # LangChain ecosystem
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "langchain-huggingface>=0.1",
    "langchain-text-splitters>=0.3",
    "langgraph>=0.4",
    # Vector store
    "langchain-qdrant>=0.2",
    "qdrant-client>=1.12",
    "fastembed>=0.4",
    # UI
    "chainlit>=2.0",
    # Voice STT + TTS
    "faster-whisper>=1.1",
    "edge-tts>=6.1",
    # Document processing
    "pypdf>=5.0",
    "pdf2image>=1.17",
    "python-docx>=1.1",
    "unstructured>=0.16",
    "openpyxl>=3.1",
    "pandas>=2.0",
    # Embeddings & Reranking
    "sentence-transformers>=2.6",
    # Web search
    "duckduckgo-search>=7.0",
    "tavily-python>=0.5",
    # MCP
    "langchain-mcp-adapters>=0.1",
    "mcp>=1.2",
    # Caching
    "cachetools>=5.3",
    # HTTP client
    "httpx>=0.28",
    # Utilities
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tiktoken>=0.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
